les lanse marchent

+++ Assets/Scripts/GameManager.cs
@@ -14,14 +14,14 @@ public class GameManager : MonoBehaviour
     public UnityEvent changeMode = new UnityEvent();
     public static GameManager instance = null;
 
-    public Transform            playerSpawnPosition;
-    public Transform            enemiesSpawnPosition;
+    public Transform playerSpawnPosition;
+    public Transform enemiesSpawnPosition;
 
-    public ViewMode            mode { get; private set; } = ViewMode.SideScroll;
+    public ViewMode mode { get; private set; } = ViewMode.SideScroll;
 
     // two lanes are between 0
-    public int      laneInterval;
-    public int      laneWidth;
+    public float laneInterval;
+    public float laneWidth;
 
     void Awake()
     {
